Move token to commons folder
